+++ internal/dataplane/kong_client.go
@@ -511,7 +511,6 @@ func (c *KongClient) sendToClient(
 	logger := c.logger.WithValues("url", client.AdminAPIClient().BaseRootURL())
 
 	deckGenParams := deckgen.GenerateDeckContentParams{
-		FormatVersion:                   config.DeckFileFormatVersion,
 		SelectorTags:                    config.FilterTags,
 		ExpressionRoutes:                config.ExpressionRoutes,
 		PluginSchemas:                   client.PluginSchemaStore(),