+++ internal/dataplane/parser/ingressclassparemeters_test.go
@@ -110,7 +110,7 @@ func TestGetIngressClassParameters(t *testing.T) {
 				Name:      testIcpName,
 			},
 			parameterSpec: defaultIcpSpec,
-			err:           store.ErrNotFound{Message: "IngressClassParameters test-icp not found"},
+			err:           store.NotFoundError{Message: "IngressClassParameters test-icp not found"},
 		},
 		{
 			name: "unmatched-name",
@@ -122,7 +122,7 @@ func TestGetIngressClassParameters(t *testing.T) {
 				Name:      "another-icp",
 			},
 			parameterSpec: defaultIcpSpec,
-			err:           store.ErrNotFound{Message: "IngressClassParameters another-icp not found"},
+			err:           store.NotFoundError{Message: "IngressClassParameters another-icp not found"},
 		},
 	}
 
@@ -153,8 +153,8 @@ func TestGetIngressClassParameters(t *testing.T) {
 			if tc.err != nil {
 				assert.EqualError(t, err, tc.err.Error())
 
-				if errors.As(tc.err, &store.ErrNotFound{}) {
-					assert.ErrorAs(t, err, &store.ErrNotFound{})
+				if errors.As(tc.err, &store.NotFoundError{}) {
+					assert.ErrorAs(t, err, &store.NotFoundError{})
 				}
 			} else {
 				assert.NoError(t, err)