+++ internal/manager/run.go
@@ -38,7 +38,6 @@ import (
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/util"
 	dataplaneutil "github.com/kong/kubernetes-ingress-controller/v2/internal/util/dataplane"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/util/kubernetes/object/status"
-	"github.com/kong/kubernetes-ingress-controller/v2/internal/versions"
 )
 
 // -----------------------------------------------------------------------------
@@ -115,14 +114,13 @@ func Run(
 	kongSemVersion := semver.Version{Major: v.Major(), Minor: v.Minor(), Patch: v.Patch()}
 
 	kongConfig := sendconfig.Config{
-		Version:               kongSemVersion,
-		InMemory:              dataplaneutil.IsDBLessMode(dbMode),
-		Concurrency:           c.Concurrency,
-		FilterTags:            c.FilterTags,
-		SkipCACertificates:    c.SkipCACertificates,
-		EnableReverseSync:     c.EnableReverseSync,
-		ExpressionRoutes:      featureGates.Enabled(featuregates.ExpressionRoutesFeature),
-		DeckFileFormatVersion: versions.DeckFileFormat(kongSemVersion),
+		Version:            kongSemVersion,
+		InMemory:           dataplaneutil.IsDBLessMode(dbMode),
+		Concurrency:        c.Concurrency,
+		FilterTags:         c.FilterTags,
+		SkipCACertificates: c.SkipCACertificates,
+		EnableReverseSync:  c.EnableReverseSync,
+		ExpressionRoutes:   featureGates.Enabled(featuregates.ExpressionRoutesFeature),
 	}
 	kongConfig.Init(ctx, setupLog, initialKongClients)