+++ internal/versions/versions.go
@@ -10,11 +10,7 @@ var (
 
 	// ExplicitRegexPathVersionCutoff is the lowest Kong version requiring the explicit "~" prefixes in regular expression paths.
 	ExplicitRegexPathVersionCutoff = semver.Version{Major: 3, Minor: 0}
-
-	// MTLSCredentialVersionCutoff is the minimum Kong version that support mTLS credentials. This is a patch version
-	// because the original version of the mTLS credential was not compatible with KIC.
-	MTLSCredentialVersionCutoff = semver.Version{Major: 2, Minor: 3, Patch: 2}
-
-	// ExpressionRouterL4Cutoff is the lowest Kong version with support of L4 proxy in expression router.
-	ExpressionRouterL4Cutoff = semver.Version{Major: 3, Minor: 4}
 )
+
+// DeckFileFormatVersion is the version of the decK file format used by KIC everywhere.
+const DeckFileFormatVersion = "3.0"