+++ test/envtest/gateway_envtest_test.go
@@ -34,7 +34,11 @@ func TestGatewayReconciliation_MoreThan100Routes(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 	gw := deployGateway(ctx, t, ctrlClient)
-	RunManager(ctx, t, envcfg, WithIngressService(gw.Namespace), WithGatewayFeatureEnabled)
+	RunManager(ctx, t, envcfg,
+		AdminAPIOptFns(),
+		WithIngressService(gw.Namespace),
+		WithGatewayFeatureEnabled,
+	)
 
 	const numOfRoutes = 120
 	createHTTPRoutes(ctx, t, ctrlClient, gw, numOfRoutes)