+++ CHANGELOG.md
@@ -116,6 +116,10 @@ Adding a new version? You'll need three changes:
 
 - No more "log.SetLogger(...) was never called..." log entry during shutdown of KIC
   [#4738](https://github.com/Kong/kubernetes-ingress-controller/pull/4738)
+- When Kong returns a flattened error related to a Kong entity, the entity's type and name
+  will be included in the message reported in `KongConfigurationApplyFailed` Kubernetes event
+  generated for it.
+  [#4813](https://github.com/Kong/kubernetes-ingress-controller/pull/4813)
 
 ### Changed